+++ R/tabularray_helpers.R
@@ -66,7 +66,7 @@ calculate_dcolumn_spec <- function(j, x) {
         "tinytable_siunitx_table_format",
         default = "table-format=-%s.%s,table-align-text-before=false,table-align-text-after=false,input-symbols={-,\\*+()}"
     )
-    num <- unlist(x@data_body[, j])
+    num <- unlist(x@body_data[, j])
 
     # empty cells
     num <- sapply(num, trimws)