+++ R/tt_tabularray.R
@@ -35,7 +35,7 @@ setMethod(
     } else {
       header <- NULL
     }
-    body <- apply(as.data.frame(x@data_body), 1, paste, collapse = " & ")
+    body <- apply(as.data.frame(x@body_data), 1, paste, collapse = " & ")
     body <- paste(body, "\\\\")
 
     # theme: booktabs