+++ R/tt_typst.R
@@ -91,12 +91,12 @@ typst_template <- function() {
 # Helper function to process table body
 typst_body <- function(x, out) {
   # Prepare body data
-  body <- apply(x@data_body, 2, function(k) paste0("[", k, "]"))
+  body <- apply(x@body_data, 2, function(k) paste0("[", k, "]"))
 
   # Apply colspan and rowspan transformations
   body <- apply_typst_spans(body, x@style)
 
-  if (nrow(x@data_body) && is.null(dim(body))) {
+  if (nrow(x@body_data) && is.null(dim(body))) {
     body <- matrix(body)
   }